+++ client/cody-shared/src/chat/recipes/file-touch.ts
@@ -20,33 +20,39 @@ import { Recipe, RecipeContext, RecipeID } from './recipe'
 /** ======================================================
  * Recipe for Generating a New File
 ====================================================== **/
-export class FileFlow implements Recipe {
-    public id: RecipeID = 'file-flow'
+export class FileTouch implements Recipe {
+    public id: RecipeID = 'file-touch'
     private workspacePath = vscode.workspace.workspaceFolders?.[0].uri
+
+    constructor(private debug: (filterLabel: string, text: string, ...args: unknown[]) => void) {}
 
     public async getInteraction(humanChatInput: string, context: RecipeContext): Promise<Interaction | null> {
         const selection = context.editor.getActiveTextEditorSelection() || context.editor.controllers?.inline.selection
         if (!selection || !this.workspacePath) {
+            await context.editor.controllers?.inline.error()
             await context.editor.showWarningMessage('Failed to start Inline Chat: empty selection.')
             return null
         }
-
         const humanInput = humanChatInput.trim() || (await this.getInstructionFromInput()).trim()
         if (!humanInput) {
+            await context.editor.controllers?.inline.error()
             await context.editor.showWarningMessage('Failed to start Inline Chat: empty input.')
             return null
         }
-
         // Get the current directory of the file that the user is currently working on
         // Create file path from selection.fileName and workspacePath
         const currentFilePath = `${this.workspacePath.fsPath}/${selection.fileName}`
         const currentDir = currentFilePath.replace(/\/[^/]+$/, '')
+        this.debug('FileTouch:currentDir', 'currentDir', currentDir)
 
         // Create new file name based on the user's input
         const newFileName = commandRegex.noTest.test(humanInput)
             ? currentFilePath.replace(/(\.[^./]+)$/, '.cody$1')
             : currentFilePath.replace(/(\.[^./]+)$/, '.test$1')
         const newFsPath = newFileName || (await this.getNewFileNameFromInput(selection.fileName, currentDir))
+        if (!newFsPath || !currentDir) {
+            return null
+        }
 
         // create vscode uri for the new file from the newFilePath which includes the workspacePath
         const fileUri = vscode.Uri.file(newFsPath)
@@ -54,6 +60,7 @@ export class FileFlow implements Recipe {
         // Create file if it doesn't exist
         workspaceEditor.createFile(fileUri, { ignoreIfExists: true })
         await vscode.workspace.applyEdit(workspaceEditor)
+        this.debug('FileTouch:workspaceEditor', 'createFile', fileUri)
 
         const truncatedText = truncateText(humanInput, MAX_HUMAN_INPUT_TOKENS)
         const MAX_RECIPE_CONTENT_TOKENS = MAX_RECIPE_INPUT_TOKENS + MAX_RECIPE_SURROUNDING_TOKENS * 2
@@ -61,7 +68,7 @@ export class FileFlow implements Recipe {
 
         // Reconstruct Cody's prompt using user's context
         // Replace placeholders in reverse order to avoid collisions if a placeholder occurs in the input
-        const prompt = FileFlow.newFilePrompt
+        const prompt = FileTouch.newFilePrompt
         const promptText = prompt
             .replace('{newFileName}', newFsPath)
             .replace('{humanInput}', truncatedText)
@@ -70,17 +77,18 @@ export class FileFlow implements Recipe {
 
         // Text display in UI fpr human that includes the selected code
         const displayText = this.getHumanDisplayText(humanInput, selection.fileName)
-
         context.responseMultiplexer.sub(
             'selection',
             new BufferedBotResponseSubscriber(async content => {
                 if (!content) {
+                    await context.editor.controllers?.inline.error()
                     await context.editor.showWarningMessage(
                         'Cody did not suggest any code updates. Please try again with a different question.'
                     )
                     return
                 }
                 await this.addContentToNewFile(workspaceEditor, fileUri, content)
+                this.debug('FileTouch:responseMultiplexer', 'BufferedBotResponseSubscriber', content)
             })
         )
 
@@ -158,25 +166,25 @@ export class FileFlow implements Recipe {
         currentDir: string
     ): Promise<ContextMessage[]> {
         const contextMessages: ContextMessage[] = []
-        // Add selected text and current file as context
-        contextMessages.push(...ChatQuestion.getEditorSelectionContext(selection))
-        // Create context messages from current directory
-        contextMessages.push(...(await FileFlow.getEditorDirContext(currentDir)))
+        // Add selected text and current file as context and create context messages from current directory
+        const selectedContext = ChatQuestion.getEditorSelectionContext(selection)
+        const currentDirContext = await FileTouch.getEditorDirContext(currentDir)
+        contextMessages.push(...selectedContext, ...currentDirContext)
+        // Create context messages from open tabs
         if (contextMessages.length < 12) {
-            // Create context messages from open tabs
-            contextMessages.push(...FileFlow.getEditorOpenTabsContext(currentDir))
+            contextMessages.push(...FileTouch.getEditorOpenTabsContext(currentDir))
         }
-
-        return contextMessages.slice(0, 12)
+        return contextMessages.slice(-12)
     }
 
     // Create Context from Current Directory of Active Document //
     public static async getEditorDirContext(currentDir: string): Promise<ContextMessage[]> {
         // get a list of files from the current directory path
         const currentDirUri = vscode.Uri.file(currentDir)
-        // Get the list of files in the current directory then filter out directories then get the first 5 files
-        const filesInDir = (await vscode.workspace.fs.readDirectory(currentDirUri)).filter(file => !file[1]).slice(0, 5)
-
+        // Get the list of files in the current directory then filter out directories and get the first 10 files
+        const filesInDir = (await vscode.workspace.fs.readDirectory(currentDirUri))
+            .filter(file => file[1] === 1)
+            .slice(0, 10)
         const contextMessages: ContextMessage[] = []
         for (const file of filesInDir) {
             // Get the context from each file
@@ -186,9 +194,7 @@ export class FileFlow implements Recipe {
             const truncatedContent = truncateText(fileContent.getText(), MAX_CURRENT_FILE_TOKENS)
             const contextMessage = getContextMessageWithResponse(
                 populateCurrentEditorContextTemplate(truncatedContent, fileName),
-                {
-                    fileName,
-                }
+                { fileName }
             )
             contextMessages.push(...contextMessage)
         }
@@ -228,7 +234,7 @@ export class FileFlow implements Recipe {
 
     // Get display text for human
     private getHumanDisplayText(humanChatInput: string, fileName: string): string {
-        return humanChatInput + FileFlow.displayPrompt + fileName
+        return humanChatInput + FileTouch.displayPrompt + fileName
     }
 
     private async getInstructionFromInput(): Promise<string> {