+++ client/cody/src/main.ts
@@ -157,7 +157,7 @@ const register = async (
         vscode.commands.registerCommand('cody.comment.add', async (comment: vscode.CommentReply) => {
             const isFixMode = /^\/f(ix)?\s/i.test(comment.text.trimStart())
             await commentController.chat(comment, isFixMode)
-            await chatProvider.executeRecipe(isFixMode ? 'fixup' : 'inline-chat', comment.text, false)
+            await chatProvider.executeRecipe('inline-chat', comment.text.trimStart(), false)
             logEvent(`CodyVSCodeExtension:inline-assist:${isFixMode ? 'fixup' : 'chat'}`)
         }),
         vscode.commands.registerCommand('cody.comment.delete', (thread: vscode.CommentThread) => {
@@ -255,7 +255,7 @@ const register = async (
             },
         }
         disposables.push(
-            vscode.commands.registerCommand('cody.recipe.file-flow', () => executeRecipe('file-flow', false))
+            vscode.commands.registerCommand('cody.recipe.file-touch', () => executeRecipe('file-touch', false))
         )
     }