+++ client/cody/src/testSetup/vscode.ts
@@ -76,6 +76,15 @@ class Range {
     }
 }
 
+class Uri {
+    public fsPath: string
+    public path: string
+    constructor(path: string) {
+        this.fsPath = path
+        this.path = path
+    }
+}
+
 class InlineCompletionItem {
     public insertText: string
     constructor(content: string) {
@@ -83,10 +92,21 @@ class InlineCompletionItem {
     }
 }
 
+// TODO(abeatrix): Implement delete and insert mocks
+class WorkspaceEdit {
+    public delete(uri: Uri, range: Range): Range {
+        return range
+    }
+    public insert(uri: Uri, position: Position, content: string): string {
+        return content
+    }
+}
+
 const vsCodeMocks = {
     Range,
     Position,
     InlineCompletionItem,
+    WorkspaceEdit,
     window: {
         showInformationMessage: () => undefined,
         showWarningMessage: () => undefined,
@@ -104,9 +124,21 @@ const vsCodeMocks = {
         getConfiguration() {
             return undefined
         },
+        openTextDocument: (uri: string) => ({
+            getText: () => 'foo\nbar\nfoo',
+            save: () => true,
+        }),
+        applyEdit: (edit: WorkspaceEdit) => true,
+        save: () => true,
     },
     ConfigurationTarget: {
         Global: undefined,
+    },
+    Uri: {
+        file: (path: string) => ({
+            fsPath: path,
+            path,
+        }),
     },
 } as const